testing Wolff for the Ising model

+++ src/marqov.h
@@ -139,7 +139,9 @@ class Marqov
 			double avgclustersize = 0;
 			for (int j=0; j<nwolff; j++)
 			{
-				const auto rdir = rnddir<RND, double, SymD>(rng);
+//				const auto rdir = rnddir<RND, double, SymD>(rng);
+				const auto rdir = rnddir<RND, int, SymD>(rng);
+
 				const int rsite = rng.i();
 //				avgclustersize += wolffstep(rsite, rdir);
 				avgclustersize += general_wolffstep(rsite, rdir);
